refactor(crypto): extract AES-256-GCM + Argon2id into nebula-crypto (ADR-0088)
Step 1 of the credential subsystem rewrite. Moves cross-cutting byte crypto out
of the nebula-credential contract crate into a new leaf cross-cutting crate:

- EncryptionKey (Argon2id KDF), EncryptedData envelope, encrypt_with_aad /
  encrypt_with_key_id / decrypt / decrypt_with_aad, CryptoError -> nebula-crypto.
- PKCE/OAuth-state helpers + serde_base64 stay in nebula-credential (they travel
  with the OAuth protocol, not generic crypto).
- nebula-credential drops the aes-gcm + argon2 deps; CryptoError is re-exported
  (it remains part of CredentialError).
- nebula-storage (EncryptionLayer/backup/key_provider) and credential-runtime
  import the primitives from nebula-crypto directly (no shim).
- CryptoError codes keep the CREDENTIAL:CRYPTO_* prefix (pure move, no behaviour change).

SEC-11: the no-AAD encrypt path stays #[cfg(test)] in nebula-crypto (uncallable
from any non-test build) -- strictly stronger than the deleted credential-side
trybuild probe, which tested a surface that no longer lives in credential.

Green: clippy -D warnings (4 crates, --all-targets) + nextest (crypto 11,
credential 251, storage 270, credential-runtime 50) + rustdoc -D warnings.
