+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
   "crates/tenancy",
   "crates/core",
   "crates/credential",
+  "crates/crypto",
   "crates/credential-builtin",
   "crates/credential-testutil",
   "crates/credential-vault",
