+++ crates/credential-runtime/src/service.rs
@@ -1298,11 +1298,12 @@ pub mod test_support {
     use nebula_core::accessor::MetricsEmitter;
     use nebula_credential::provider::LeaseEvent;
     use nebula_credential::store::StoreError;
-    use nebula_credential::{CredentialEvent, CredentialId, CredentialRegistry, EncryptionKey};
+    use nebula_credential::{CredentialEvent, CredentialId, CredentialRegistry};
     use nebula_credential_builtin::{
         BearerTokenCredential, SharedKeyCredential, SigningKeyCredential, register_builtins,
     };
     use nebula_credential_testutil::InMemoryPendingStore;
+    use nebula_crypto::EncryptionKey;
     use nebula_engine::credential::LeaseLifecycleConfig;
     use nebula_eventbus::EventBus;
     use nebula_storage::credential::{
