+++ crates/credential/Cargo.toml
@@ -19,9 +19,8 @@ repository.workspace = true
 tokio = { workspace = true, features = ["time", "macros", "rt"] }
 tokio-util = { workspace = true }
 
-# Crypto and security
-aes-gcm = "0.10"
-argon2 = "0.5"
+# Crypto and security — AES-256-GCM + Argon2id live in nebula-crypto (ADR-0088).
+nebula-crypto = { path = "../crypto" }
 base64 = { workspace = true }
 rand = { workspace = true }
 sha2 = { workspace = true }
