--- crates/credential/tests/probes/encrypt_no_aad_removed.stderr
@@ -1,29 +1,0 @@
-error[E0432]: unresolved import `nebula_credential::encrypt`
-  --> tests/probes/encrypt_no_aad_removed.rs:26:9
-   |
-26 |     use nebula_credential::encrypt;
-   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^ no `encrypt` in the root
-   |
-help: a similar name exists in the module
-   |
-26 -     use nebula_credential::encrypt;
-26 +     use nebula_credential::decrypt;
-   |
-
-error[E0425]: cannot find function `encrypt` in `nebula_credential::secrets::crypto`
-  --> tests/probes/encrypt_no_aad_removed.rs:21:49
-   |
-21 |     let _ = nebula_credential::secrets::crypto::encrypt(&key, plaintext);
-   |                                                 ^^^^^^^ not found in `nebula_credential::secrets::crypto`
-
-error[E0603]: module `crypto` is private
-  --> tests/probes/encrypt_no_aad_removed.rs:21:41
-   |
-21 |     let _ = nebula_credential::secrets::crypto::encrypt(&key, plaintext);
-   |                                         ^^^^^^ private module
-   |
-note: the module `crypto` is defined here
-  --> src/secrets/mod.rs
-   |
-   | mod crypto;
-   | ^^^^^^^^^^
