+++ crates/crypto/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "nebula-crypto"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+keywords.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+# AES-256-GCM authenticated encryption + Argon2id KDF — the only reason this
+# crate exists. Pinned to the same versions nebula-credential carried before the
+# ADR-0088 extract.
+aes-gcm = "0.10"
+argon2 = "0.5"
+zeroize = { version = "1.8.2", features = ["zeroize_derive"] }
+rand = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+
+# CryptoError derives the cross-cutting Classify taxonomy (needs the derive feature).
+nebula-error = { workspace = true, features = ["derive"] }
+thiserror = { workspace = true }
+
+[lints]
+workspace = true
